mamma mia che frustrazione
potremmo esserci
corretti alcuni errori
continua il problema del refresh che non porta i dati su schermo
