+++ project_mobile/pubspec.yaml
@@ -30,14 +30,14 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  shared_preferences: ^2.0.8
+  shared_preferences: ^2.2.2
   table_calendar: ^3.1.0
   google_fonts: ^6.1.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  fl_chart: ^0.60.0
+  fl_chart: ^0.67.0
   firebase_core: ^2.27.1
   firebase_auth: ^4.17.9
